Avoid abi changes

+++ src/bson/bson-string.c
@@ -63,7 +63,7 @@ bson_string_new (const char *str) /* IN */
    ret->alloc = ret->len + 1;
 
    if (!bson_is_power_of_two (ret->alloc)) {
-      ret->alloc = bson_next_power_of_two ((size_t)ret->alloc);
+      ret->alloc = (uint32_t)bson_next_power_of_two ((size_t)ret->alloc);
    }
 
    BSON_ASSERT (ret->alloc >= 1);
@@ -151,7 +151,7 @@ bson_string_append (bson_string_t *string, /* IN */
    if ((string->alloc - string->len - 1) < len) {
       string->alloc += len;
       if (!bson_is_power_of_two (string->alloc)) {
-         string->alloc = bson_next_power_of_two ((size_t)string->alloc);
+         string->alloc = (uint32_t)bson_next_power_of_two ((size_t)string->alloc);
       }
       string->str = bson_realloc (string->str, string->alloc);
    }
@@ -294,7 +294,7 @@ void
 bson_string_truncate (bson_string_t *string, /* IN */
                       uint32_t       len)    /* IN */
 {
-   size_t alloc;
+   uint32_t alloc;
 
    bson_return_if_fail (string);
    bson_return_if_fail (len < INT_MAX);
@@ -306,7 +306,7 @@ bson_string_truncate (bson_string_t *string, /* IN */
    }
 
    if (!bson_is_power_of_two (alloc)) {
-      alloc = bson_next_power_of_two ((size_t)alloc);
+      alloc = (uint32_t)bson_next_power_of_two ((size_t)alloc);
    }
 
    string->str = bson_realloc (string->str, alloc);
